requirements.txt
uvicorn[standard]
fastapi
pydantic
requests
python-dotenv
mysql-connector-python
jinja2
python-multipart
redis